+++ ThanosGame/weapons/Rifle.java
@@ -7,7 +7,7 @@ import resources.AudioSaves;
 import java.util.LinkedList;
 
 public class Rifle extends Weapon {
-    public Rifle(LinkedList<Projectile> worldProjectiles, Personnage owner, int damage) {
+    public Rifle(LinkedList<Projectile> worldProjectiles, Personnage owner, double damage) {
         super(worldProjectiles, owner, damage, 5, 1500);
     }
