+++ ThanosGame/weapons/Spear.java
@@ -6,7 +6,7 @@ import javafx.geometry.Point2D;
 import java.util.LinkedList;
 
 public class Spear extends Weapon {
-    public Spear(LinkedList<Projectile> worldProjectiles, Personnage owner, int damage) {
+    public Spear(LinkedList<Projectile> worldProjectiles, Personnage owner, double damage) {
         super(worldProjectiles, owner, damage, 5, 500);
     }
